fix: fix unbound ns erro
